requirements-gradio.txt
gradio==4.44.0
transformers==4.35.0
torch==2.1.0
Pillow>=10.2.0
